Style: styled inital-info and authenticating component
