+++ frontend/src/components/authenticating/Authenticating.tsx
@@ -1,0 +1,10 @@
+export function Authenticating() {
+    return (
+        <div style={{ display: 'flex', justifyContent: 'space-around', alignItems: 'center', height: '100vh', width: '100%'  }}>
+            <h1>
+                Please wait! while we log you in...
+            </h1>
+        </div>
+
+    )
+}
